used-car-price-system/Back-End/requirements.txt
Flask==2.0.3
Flask-Cors==4.0.0
mysql-connector-python==8.0.26
python-dotenv==0.19.0
scikit-learn==0.24.2
pandas==1.3.2
numpy==1.21.2
joblib==1.0.1
matplotlib==3.4.3